lambda_function/requirements.txt
boto3==1.34.140
botocore==1.34.140
jmespath==1.0.1
numpy==2.0.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.2
six==1.16.0
tzdata==2024.1
urllib3==1.26.19
openpyxl==3.1.3